requirements.txt
streamlit
pandas
numpy
matplotlib
pandas_datareader
sklearn
tensorflow
keras
Pillow
tweepy
regex
textblob
